Harden login resilience against network errors (v1.8.8)
- Add error categorization to distinguish transient network errors from auth failures
- Implement retry logic with exponential backoff (up to 3 retries)
- Only clear credentials on actual auth failures, not transient network issues
- Update iOS, VisionOS, and macOS error handling to preserve credentials during network errors
- Extract widget formatting to QueryFluxWidgetExtensions.swift to reduce file size
- Verify SwiftLint compliance

This ensures users won't be logged out due to temporary network issues while still properly handling authentication failures.
